+++ test/controllers/topics_controller_test.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require "test_helper"
 
 # Tests to assure that community members can create and view forum topics.
-class TopicsControllerTest < ActionController::TestCase
+class TopicsControllerTest < ActionDispatch::IntegrationTest
   setup do
     @topic = topics(:one)
     @regular_user = users(:user_1)
@@ -11,83 +11,109 @@ class TopicsControllerTest < ActionController::TestCase
 
   def topic_params
     {
-      title: 'Topic Title',
-      slug: 'topic-title',
-      description: 'This is the my new forum topic.',
-      pinned: '1'
+      title: "New Topic Title",
+      slug: "new-topic-title",
+      description: "This is the my new forum topic.",
+      pinned: "1"
     }
   end
 
-  test 'should get index' do
+  test "should subscribe to notifications" do
+    login(users(:valid))
+    post subscription_topic_url(@topic, format: "js"), params: { notify: "1" }
+    assert_not_nil assigns(:topic)
+    assert_equal true, assigns(:topic).subscribed?(users(:valid))
+    assert_template "subscription"
+    assert_response :success
+  end
+
+  test "should unsubscribe from notifications" do
+    login(users(:valid))
+    post subscription_topic_url(@topic, format: "js"), params: { notify: "0" }
+    assert_not_nil assigns(:topic)
+    assert_equal false, assigns(:topic).subscribed?(users(:valid))
+    assert_template "subscription"
+    assert_response :success
+  end
+
+  test "should not subscribe for anonymous user" do
+    post subscription_topic_url(@topic, format: "js"), params: { notify: "1" }
+    assert_template nil
+    assert_response :unauthorized
+  end
+
+  test "should get index" do
     login(@regular_user)
-    get :index
+    get topics_url
     assert_response :success
     assert_not_nil assigns(:topics)
   end
 
-  test 'should get new' do
+  test "should get new" do
     login(@regular_user)
-    get :new
+    get new_topic_url
     assert_response :success
   end
 
-  test 'should create topic' do
+  test "should create topic" do
     login(@regular_user)
-    assert_difference('Topic.count') do
-      post :create, params: { topic: topic_params }
+    assert_difference("Topic.count") do
+      post topics_url, params: { topic: topic_params }
     end
     assert_not_nil assigns(:topic)
-    assert_equal 'Topic Title', assigns(:topic).title
-    assert_equal 'This is the my new forum topic.', assigns(:topic).replies.first.description
+    assert_equal "New Topic Title", assigns(:topic).title
+    assert_equal "This is the my new forum topic.", assigns(:topic).replies.first.description
     assert_equal users(:user_1), assigns(:topic).user
     assert_redirected_to assigns(:topic)
   end
 
-  test 'should not create topic with blank title' do
+  test "should not create topic with blank title" do
     login(@regular_user)
-    assert_difference('Topic.count', 0) do
-      post :create, params: { topic: topic_params.merge(title: '') }
+    assert_difference("Topic.count", 0) do
+      post topics_url, params: { topic: topic_params.merge(title: "") }
     end
-    assert_template 'new'
+    assert_template "new"
     assert_response :success
   end
 
-  test 'should shadow ban new spammer' do
+  test "should shadow ban new spammer" do
     login(users(:new_spammer))
-    assert_difference('Topic.count') do
-      post :create, params: { topic: { title: 'http://www.example.com', slug: 'http-www-example-com', description: 'http://www.example.com' } }
+    assert_difference("Topic.count") do
+      post topics_url, params: {
+        topic: {
+          title: "http://www.example.com",
+          slug: "http-www-example-com",
+          description: "http://www.example.com" }
+        }
     end
     assert_equal true, assigns(:topic).user.shadow_banned
     assert_redirected_to assigns(:topic)
   end
 
-  test 'should not shadow ban verified user' do
+  test "should not shadow ban verified user" do
     login(users(:verified_user))
-    assert_difference('Topic.count') do
-      post :create, params: { topic: { title: 'http://www.example.com', slug: 'http-www-example-com', description: 'http://www.example.com' } }
+    assert_difference("Topic.count") do
+      post topics_url, params: {
+        topic: {
+          title: "http://www.example.com",
+          slug: "http-www-example-com",
+          description: "http://www.example.com" }
+        }
     end
     assert_equal false, assigns(:topic).user.shadow_banned
     assert_redirected_to assigns(:topic)
   end
 
-  # test 'should not create topic for shadow banned user' do
-  #   login(users(:shadow_banned))
-  #   assert_difference('Topic.count', 0) do
-  #     post :create, params: { topic: topic_params.merge(title: '') }
-  #   end
-  #   assert_redirected_to topics_path
-  # end
-
-  test 'should show topic' do
+  test "should show topic" do
     login(@regular_user)
-    get :show, params: { id: @topic }
+    get topic_url(@topic)
     @topic.reload
     assert_equal 1, @topic.view_count
     assert_response :success
   end
 
   test "should increase topic view count as public viewer" do
-    get :show, params: { id: @topic }
+    get topic_url(@topic)
     @topic.reload
     assert_equal 1, @topic.view_count
     assert_response :success
@@ -95,35 +121,35 @@ class TopicsControllerTest < ActionController::TestCase
 
   test "should not increase topic view count as spammer" do
     login(users(:shadow_banned))
-    get :show, params: { id: @topic }
+    get topic_url(@topic)
     @topic.reload
     assert_equal 0, @topic.view_count
     assert_response :success
   end
 
-  test 'should get edit' do
+  test "should get edit" do
     login(@regular_user)
-    get :edit, params: { id: @topic }
+    get edit_topic_url(@topic)
     assert_response :success
   end
 
-  test 'should update topic' do
+  test "should update topic" do
     login(@regular_user)
-    patch :update, params: { id: @topic, topic: topic_params }
+    patch topic_url(@topic), params: { topic: topic_params.merge(title: "Updated Topic Title") }
     assert_redirected_to assigns(:topic)
   end
 
-  test 'should not update topic with blank title' do
+  test "should not update topic with blank title" do
     login(@regular_user)
-    patch :update, params: { id: @topic, topic: topic_params.merge(title: '') }
-    assert_template 'edit'
+    patch topic_url(@topic), params: { topic: topic_params.merge(title: "") }
+    assert_template "edit"
     assert_response :success
   end
 
-  test 'should destroy topic' do
+  test "should destroy topic" do
     login(@regular_user)
-    assert_difference('Topic.current.count', -1) do
-      delete :destroy, params: { id: @topic }
+    assert_difference("Topic.current.count", -1) do
+      delete topic_url(@topic)
     end
     assert_redirected_to topics_path
   end